requirements.txt
Pillow==9.5.0
torch==1.13.1
torchvision==0.14.1